rename : input, button 파일 common 폴더로 이동 #20
